server/requirements.txt
bottle
pyyaml
